backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.3
python-dotenv==1.0.0
groq==0.4.1
pydantic==2.5.0
